+++ SampleCode/StdDriver/USBH_MassStorage/diskio.c
@@ -68,11 +68,11 @@ DRESULT disk_read
 
     //printf("disk_read - drv:%d, sec:%d, cnt:%d, buff:0x%x\n", pdrv, sector, count, (uint32_t)buff);
 
-    ret = usbh_umas_read(pdrv, sector, count, buff);
+    ret = usbh_umas_read(pdrv, sector, (int)count, buff);
     if(ret != UMAS_OK)
     {
         usbh_umas_reset_disk(pdrv);
-        ret = usbh_umas_read(pdrv, sector, count, buff);
+        ret = usbh_umas_read(pdrv, sector, (int)count, buff);
     }
 
     if(ret == UMAS_OK)
@@ -106,11 +106,11 @@ DRESULT disk_write
 
     //printf("disk_write - drv:%d, sec:%d, cnt:%d, buff:0x%x\n", pdrv, sector, count, (uint32_t)buff);
 
-    ret = usbh_umas_write(pdrv, sector, count, (uint8_t *)buff);
+    ret = usbh_umas_write(pdrv, sector, (int)count, (uint8_t *)(uint32_t)buff);
     if(ret != UMAS_OK)
     {
         usbh_umas_reset_disk(pdrv);
-        ret = usbh_umas_write(pdrv, sector, count, (uint8_t *)buff);
+        ret = usbh_umas_write(pdrv, sector, (int)count, (uint8_t *)(uint32_t)buff);
     }
 
     if(ret == UMAS_OK)
